Rename google-chart.js to GoogleChart.js.
