+++ assets/js/modules/analytics/components/dashboard/AnalyticsDashboardWidgetSiteStats.js
@@ -28,7 +28,7 @@ import { Component } from '@wordpress/element';
 import {
 	getTimeInSeconds,
 } from '../../../../util';
-import GoogleChart from '../../../../components/google-chart';
+import GoogleChart from '../../../../components/GoogleChart';
 import withData from '../../../../components/higherorder/withdata';
 import { TYPE_MODULES } from '../../../../components/data';
 import { extractAnalyticsDashboardData, siteAnalyticsReportDataDefaults } from '../../util';
