+++ assets/js/modules/analytics/components/dashboard/LegacyDashboardAcquisitionPieChart.js
@@ -34,7 +34,7 @@ import Data from 'googlesitekit-data';
 import { STORE_NAME } from '../../datastore/constants';
 import { STORE_NAME as CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
 import { getTimeInSeconds } from '../../../../util';
-import GoogleChart from '../../../../components/google-chart';
+import GoogleChart from '../../../../components/GoogleChart';
 import withData from '../../../../components/higherorder/withdata';
 import { TYPE_MODULES } from '../../../../components/data';
 import Link from '../../../../components/Link';
